World coastline at medium-scale (1:50M) resolution

coastlineWorldMedum is analogous to coastlineWorld in the oce package, but at higher resolution, i.e. showing more detail. It is suitable for basin-scale plots; for sub-basin work, try coastlineWorldFine.

data(coastlineWorldMedium, package="ocedata")

Details

The data for coastlineWorldMedium were downloaded from shapefile ne_50m_admin_0_countries.shp, stored at https://www.naturalearthdata.com.

The documentation for coastline-class in oce explains the structure of coastline objects and discusses functions that deal with them.

See also

Examples

# Regional plot of the Maritime provinces of Canada.
data(coastlineWorldMedium, package="ocedata")
if (FALSE) {
library(oce)
plot(coastlineWorldMedium, clongitude=-63, clatitude=45.5, span=600)
}